ML/train-api/requirement.txt
Flask
tensorFlow
scikit-learn
gunicorn
numpy
Flask-HTTPAuth
"cloud-sql-python-connector[pymysql]"
google-cloud-sql-connector
sqlalchemy
google-api-python-client
google-cloud-storage